requirements.txt
Django==5.0.10
djangorestframework==3.15.2
Django==5.0.10
djangorestframework==3.15.2
django-cors-headers==4.3.1
mssql-django==1.5
mssql-django==1.5
Pillow==10.3.0
google-generativeai
PyJWT
requests
gunicorn
pandas
openpyxl
django-storages
azure-storage-blob